Merge with base

+++ ccstruct/coutln.cpp
@@ -1,5 +1,5 @@
 /**********************************************************************
- * File:        coutln.c  (Formerly coutline.c)
+ * File:        coutln.cpp  (Formerly coutline.c)
  * Description: Code for the C_OUTLINE class.
  * Author:                  Ray Smith
  * Created:                 Mon Oct 07 16:01:57 BST 1991
@@ -603,7 +603,7 @@ void C_OUTLINE::move(const ICOORD vec) {
  */
 bool C_OUTLINE::IsLegallyNested() const {
   if (stepcount == 0) return true;
-  int parent_area = outer_area();
+  int64_t parent_area = outer_area();
   // We aren't going to modify the list, or its contents, but there is
   // no const iterator.
   C_OUTLINE_IT child_it(const_cast<C_OUTLINE_LIST*>(&children));
@@ -652,22 +652,10 @@ static void ComputeGradient(const l_uint32* data, int wpl,
                             int x, int y, int width, int height,
                             ICOORD* gradient) {
   const l_uint32* line = data + y * wpl;
-  int pix_x_y =
-      x < width && y < height
-          ? GET_DATA_BYTE(line, x)
-          : 255;
-  int pix_x_prevy =
-      x < width && y > 0
-          ? GET_DATA_BYTE(line - wpl, x)
-          : 255;
-  int pix_prevx_prevy =
-      x > 0 && y > 0
-          ? GET_DATA_BYTE(line - wpl, x - 1)
-          : 255;
-  int pix_prevx_y =
-      x > 0 && y < height
-          ? GET_DATA_BYTE(line, x - 1)
-          : 255;
+  int pix_x_y = x < width && y < height ? GET_DATA_BYTE(line, x) : 255;
+  int pix_x_prevy = x < width && y > 0 ? GET_DATA_BYTE(line - wpl, x) : 255;
+  int pix_prevx_prevy = x > 0 && y > 0 ? GET_DATA_BYTE(line - wpl, x - 1) : 255;
+  int pix_prevx_y = x > 0 && y < height ? GET_DATA_BYTE(line, x - 1) : 255;
   gradient->set_x(pix_x_y + pix_x_prevy - (pix_prevx_y + pix_prevx_prevy));
   gradient->set_y(pix_x_prevy + pix_prevx_prevy - (pix_x_y + pix_prevx_y));
 }
